added javadocs

+++ Task_1_3_1/src/test/java/ru/nsu/krasnikov/SubStringSearchTest.java
@@ -3,17 +3,19 @@ package ru.nsu.krasnikov;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.util.List;
-
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+/**
+ * Tests for searching substring in file.
+ */
 public class SubStringSearchTest {
     @Test
     public void test1() throws IOException {
         SubStringSearch res1 = new SubStringSearch(
                 ClassLoader.getSystemResource("test1.txt").getFile(), "ab");
-        List<Integer> answer = res1.findIndexes();
-        Integer[] correct = new Integer[]{1, 3, 6};
+        List<Long> answer = res1.findIndexes();
+        Long[] correct = new Long[]{1L, 3L, 6L};
         Assertions.assertArrayEquals(answer.toArray(), correct);
     }
 
@@ -21,8 +23,8 @@ public class SubStringSearchTest {
     public void test2() throws IOException {
         SubStringSearch res1 = new SubStringSearch(
                 ClassLoader.getSystemResource("test2.txt").getFile(), "abaaba");
-        List<Integer> answer = res1.findIndexes();
-        Integer[] correct = new Integer[]{11, 14};
+        List<Long> answer = res1.findIndexes();
+        Long[] correct = new Long[]{11L, 14L};
         Assertions.assertArrayEquals(answer.toArray(), correct);
     }
 
@@ -30,8 +32,8 @@ public class SubStringSearchTest {
     public void test3() throws IOException {
         SubStringSearch res1 = new SubStringSearch(
                 ClassLoader.getSystemResource("test3.txt").getFile(), "a");
-        List<Integer> answer = res1.findIndexes();
-        Integer[] correct = new Integer[]{0, 1, 2, 3, 5, 6, 7};
+        List<Long> answer = res1.findIndexes();
+        Long[] correct = new Long[]{0L, 1L, 2L, 3L, 5L, 6L, 7L};
         Assertions.assertArrayEquals(answer.toArray(), correct);
     }
 
@@ -39,8 +41,8 @@ public class SubStringSearchTest {
     public void test4() throws IOException {
         SubStringSearch res1 = new SubStringSearch(
                 ClassLoader.getSystemResource("test4.txt").getFile(), "aa");
-        List<Integer> answer = res1.findIndexes();
-        Integer[] correct = new Integer[]{0, 1, 2, 5, 6};
+        List<Long> answer = res1.findIndexes();
+        Long[] correct = new Long[]{0L, 1L, 2L, 5L, 6L};
         Assertions.assertArrayEquals(answer.toArray(), correct);
     }
 
@@ -48,8 +50,8 @@ public class SubStringSearchTest {
     public void test5() throws IOException {
         SubStringSearch res1 = new SubStringSearch(
                 ClassLoader.getSystemResource("test5.txt").getFile(), "seven");
-        List<Integer> answer = res1.findIndexes();
-        Integer[] correct = new Integer[]{28, 97, 132, 140, 152, 160, 166, 180, 192, 206, 216, 230, 238};
+        List<Long> answer = res1.findIndexes();
+        Long[] correct = new Long[]{28L, 97L, 132L, 140L, 152L, 160L, 166L, 180L, 192L, 206L, 216L, 230L, 238L};
         Assertions.assertArrayEquals(answer.toArray(), correct);
     }
 
@@ -57,8 +59,8 @@ public class SubStringSearchTest {
     public void test6() throws IOException {
         SubStringSearch res1 = new SubStringSearch(
                 ClassLoader.getSystemResource("test6.txt").getFile(), "123456789");
-        List<Integer> answer = res1.findIndexes();
-        Integer[] correct = new Integer[]{0};
+        List<Long> answer = res1.findIndexes();
+        Long[] correct = new Long[]{0L};
         Assertions.assertArrayEquals(answer.toArray(), correct);
     }
 
@@ -66,8 +68,8 @@ public class SubStringSearchTest {
     public void test7() throws IOException {
         SubStringSearch res1 = new SubStringSearch(
                 ClassLoader.getSystemResource("test7.txt").getFile(), "d");
-        List<Integer> answer = res1.findIndexes();
-        Integer[] correct = new Integer[]{};
+        List<Long> answer = res1.findIndexes();
+        Long[] correct = new Long[]{};
         Assertions.assertArrayEquals(answer.toArray(), correct);
     }
